+++ src/app/tokens/tests/test_api.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Krypto Explorer - Token API - Tests
-"""