+++ src/app/tokens/tests/test_crud.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Krypto Explorer - Token CRUD Operations - Tests
-
-"""