+++ src/app/tokens/tests/test_model.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Krypto Explorer - Token Models - Tests
-"""