+++ src/app/tokens/tests/test_schema.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Krypto Explorer Schema - Token Schemas - Tests
-"""