+++ src/app/tokens/tests/test_service.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Krypto Explorer CLI - Token Services - Tests
-"""