+++ src/qrypt/tokens/schema.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 """
-Krypto Explorer Schema - Token Schemas
+Qrypto Schema - Token Schemas
 
-This module contains the schema for the Krypto Explorer application.
+This module contains the schema for the Qrypto application.
 
 It defines the data models and their relationships using SQLAlchemy and Pydantic.
 """