+++ src/qrypt/tokens/tests/test_crud.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+"""
+Qrypto - Token CRUD Operations - Tests
+
+"""