+++ src/qrypt/tokens/tests/test_schema.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""
+Qrypto Schema - Token Schemas - Tests
+"""