+++ src/qrypt/tokens/tests/test_service.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+"""
+Qrypto CLI - Token Services - Tests
+
+"""